Should be **at least 4 seconds** faster. Although that depends on the user's specs.

I got from 16.25s to 10.23s and with all Installer Addons it was 18s to 12.28s

My specs:
![image](https://user-images.githubusercontent.com/31479541/42128589-e14ef284-7ca6-11e8-9447-b2ad0150ee6e.png)

I've also tested everything that I've multithreaded etc, they all work.